Resolves #1. - Parser extracted - Added second Parse method for easy usage
